restructure ui-utils
