+++ packages/ui-utils/src/stream-string.ts
@@ -1,0 +1,10 @@
+import { StreamStringPrefixes } from './stream-parts';
+
+export const isStreamStringEqualToType = (
+  type: keyof typeof StreamStringPrefixes,
+  value: string,
+): value is StreamString =>
+  value.startsWith(`${StreamStringPrefixes[type]}:`) && value.endsWith('\n');
+
+export type StreamString =
+  `${(typeof StreamStringPrefixes)[keyof typeof StreamStringPrefixes]}:${string}\n`;
